+++ src/Sulu/Bundle/MediaBundle/Domain/Event/CollectionModifiedEvent.php
@@ -99,9 +99,7 @@ class CollectionModifiedEvent extends DomainEvent
         $meta = $this->collection->getDefaultMeta();
         foreach ($this->collection->getMeta() as $collectionMeta) {
             if ($collectionMeta->getLocale() === $this->locale) {
-                $meta = $collectionMeta;
-
-                break;
+                return $collectionMeta;
             }
         }
