+++ src/Sulu/Bundle/MediaBundle/Domain/Event/MediaCropRemovedEvent.php
@@ -44,6 +44,11 @@ class MediaCropRemovedEvent extends DomainEvent
         return $this->media;
     }
 
+    public function getFormatKey(): string
+    {
+        return $this->formatKey;
+    }
+
     public function getEventType(): string
     {
         return 'crop_removed';
